+++ src/event.js
@@ -1,95 +1,21 @@
 
-var iris = {};
-
-// Expose iris to the global object
-window.iris = iris;
-
 
 (function ($) {
 
-    // static object to store all app callbacks
-    var _events;
-    
     function _init() {
-
-        _events = {};
+        iris.events = {};
 
         iris.on("iris-reset", _init);
     }
 
-    iris.on = function (p_eventName, f_func) {
-        
-        if ( ! $.isFunction(f_func) ) {
-            throw "invalid function";
-        }
-
-        if ( !_events.hasOwnProperty(p_eventName) ) {
-            _events[p_eventName] = [];
-        }
-
-        var callbacks = _events[p_eventName];
-        var index = $.inArray(f_func, callbacks);
-        if ( index === -1 ) {
-            callbacks.push(f_func);
-        }
-
-    };
-
-    iris.off = function (p_eventName, f_func){
-        if ( _events.hasOwnProperty(p_eventName) ){
-            if (f_func !== undefined) {
-                
-                if ( ! $.isFunction(f_func) ) {
-                    throw "invalid function";
-                }
-
-                var index = $.inArray(f_func, _events[p_eventName]);
-                if ( index !== -1 ) {
-                    _events[p_eventName].splice(index, 1);
-                }
-
-            } else {
-                delete _events[p_eventName];
-            }
-        }
-    };
-
-    iris.notify = function (p_eventName, p_data){
-        if ( p_eventName === undefined ) {
-            throw "event name undefined";
-        }
-        
-        if ( _events[p_eventName] ) {
-            var callbacks = _events[p_eventName];
-            for ( var i=0; i < callbacks.length; i++ ) {
-                callbacks[i](p_data);
-            }
-        }
-    };
-
-    iris.destroyEvents = function (p_eventName, p_callbacks) {
-        // Create an array copy, to prevent concurrent modification of _events[p_eventName] array.
-        // This occur if an event that destroy uis is notified
-        var callbacks = _events[p_eventName].concat([]);
-        for ( var i=0; i < p_callbacks.length; i++ ) {
-            var index = $.inArray(p_callbacks[i], callbacks);
-            if ( index !== -1 ) {
-                callbacks.splice(index, 1);
-            }
-        }
-        _events[p_eventName] = callbacks;
-    };
-
-
-    iris.Event = function () {
-
+    var Event = function () {
         this.events = {}; // { "event1" : [f1, f2], "event2" : [f3, f4, f5, f6] }
-
     };
 
-    var eventPrototype = iris.Event.prototype;
+    var eventPrototype = Event.prototype;
 
     eventPrototype.on = function (p_eventName, f_func) {
+
         if ( ! $.isFunction(f_func) ) {
             throw "invalid function";
         }
@@ -102,7 +28,6 @@ window.iris = iris;
         var index = $.inArray(f_func, callbacks);
         if ( index === -1 ) {
             callbacks.push(f_func);
-            iris.on(p_eventName, f_func);
         }
 
     };
@@ -118,27 +43,42 @@ window.iris = iris;
         if ( callbacks ) {
 
             if (f_func !== undefined) {
-                var index = $.inArray(f_func, callbacks);
 
+                var index = $.inArray(f_func, callbacks);
                 if ( index !== -1 ) {
                     callbacks.splice(index, 1);
-                    iris.off(p_eventName, f_func);
                 }
 
             } else {
-
-                for (var i = 0; i < callbacks.length; i++ ) {
-                    iris.off(p_eventName, callbacks[i]);
-                }
-
-                this.events[p_eventName] = {};
+                delete this.events[p_eventName];
             }
         }
     };
 
-    eventPrototype.notify = function (p_eventName, p_data){
-        iris.notify(p_eventName, p_data);
+    eventPrototype.notify = function (p_eventName, p_data) {
+        if ( p_eventName === undefined ) {
+            throw "event name undefined";
+        }
+        
+        var callbacks = this.events[p_eventName];
+        if ( callbacks ) {
+            for ( var i = 0; i < callbacks.length; i++ ) {
+                callbacks[i].call(this, p_data);
+            }
+        }
     };
+
+
+    var Iris = function() {
+        Event.call(this);
+    };
+    Iris.prototype = new Event();
+    var iris = new Iris();
+
+    //
+    // Public
+    //
+    iris.Event = Event;
 
     //
     // Iris custom events
@@ -147,6 +87,8 @@ window.iris = iris;
     iris.AFTER_NAVIGATION = "iris_after_navigation";
     iris.RESOURCE_ERROR = "iris_resource_error";
     iris.SCREEN_NOT_FOUND = "iris_screen_not_found";
+
+    window.iris = iris;
     
     _init();