+++ RELEASES.md
@@ -5,7 +5,10 @@ _Creation date: 2012-01-12_
 [version] date -> authors
   * upd|fix|new|dep|rmv - description
 
-[0.5.7-SNAPSHOT] 2014-XX-XX
+[0.6.0-SNAPSHOT] 2014-XX-XX
+ * [new] iris.Data
+ * [new] Resource & Component events
+ * [upd] self.on, self.off use local events (no backward compability)
 
 
 [0.5.6] 2014-01-23